git-svn-id: http://code.taobao.org/svn/tair/trunk@1011 ed5bfd34-b61d-4ad5-b956-f82ce90ca771

+++ src/storage/ldb/leveldb/db/skiplist.h
@@ -105,7 +105,8 @@ class SkipList {
   port::AtomicPointer max_height_;   // Height of the entire list
 
   inline int GetMaxHeight() const {
-    return reinterpret_cast<intptr_t>(max_height_.NoBarrier_Load());
+    return static_cast<int>(
+        reinterpret_cast<intptr_t>(max_height_.NoBarrier_Load()));
   }
 
   // Read/written only by Insert().
@@ -375,4 +376,4 @@ bool SkipList<Key,Comparator>::Contains(const Key& key) const {
   }
 }
 
-}
+}  // namespace leveldb
